server/requirements.txt
blinker==1.8.2
click==8.1.8
Flask==3.0.3
Flask-Cors==5.0.0
flask-marshmallow==1.2.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
marshmallow==3.22.0
packaging==25.0
python-dotenv==1.0.1
SQLAlchemy==2.0.43
typing_extensions==4.13.2
Werkzeug==3.0.6
zipp==3.20.2
